:recycle: refactor: broke toolbar out into multiple components
